+++ lib/service_actor/result.rb
@@ -104,7 +104,7 @@ class ServiceActor::Result < BasicObject
   attr_reader :data
 
   # Key `_default_output` is an internal datum used by actor class
-  # method `.output`. Don't expose it with the rest of the result.
+  # method `.valuable`. Don't expose it with the rest of the result.
   def filter_default_output(h)
     # using `filter` instead of `except` to maintain Ruby 2.7 compatibility
     # update once support for 2.7 is dropped